Fix "reference before assignment" error when logging ignored entities
This only causes an error when the earlier `for entity in ordered_entities:` does not create a variable named `entity` (i.e. when there are not yet any replaceable entities in the project). However even in the case where it doesn't throw an error, it will log erroneous information (because it is referencing `entity` rather than `db_entity` which is the object being considered).