Update pg_materialized_view.py
Thanks for the fast reply.
Ive implemented the usage of List(Index) instead of my own-generated List. Problem there is, I dont know how to get the import Index in the autogenerated migrationfile. I have to speak with my colleagues about that. Im learning/using python since start of this year, so it maybe isnt the best pythonic way Ive implemented it.
cls.from_database seems ok with it. As long as I dont change something in the view definition the autogenerate dont create a replace statement. Problem is, that if you would need a replace (ie change of defined columns for the MatView) the downgrade dont know the indexes.
Look into cls.get_required_migration_op would need a bit of time.